requirements.txt
python-telegram-bot
transformers
torch
gunicorn